requirements.txt
openai
fastapi
uvicorn[standard]
pydantic
httpx
pillow
python-multipart

